refactor: one less useEffect
