+++ frontend/src/components/Board/RegularBoard/index.tsx
@@ -1,16 +1,12 @@
 import React, { useMemo, useState } from 'react';
-import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { useRecoilValue } from 'recoil';
 
 import { Container } from '@/styles/pages/boards/board.styles';
 
 import DragDropArea from '@/components/Board/DragDropArea';
 import LoadingPage from '@/components/loadings/LoadingPage';
 import Flex from '@/components/Primitives/Flex';
-import {
-  boardInfoState,
-  deletedColumnsState,
-  editColumnsState,
-} from '@/store/board/atoms/board.atom';
+import { boardInfoState } from '@/store/board/atoms/board.atom';
 import { BoardUserRoles } from '@/utils/enums/board.user.roles';
 import Button from '@/components/Primitives/Button';
 import Icon from '@/components/icons/Icon';
@@ -29,15 +25,6 @@ const RegularBoard = ({ socketId }: RegularBoardProps) => {
 
   // Recoil States
   const { board } = useRecoilValue(boardInfoState);
-  const setEditColumns = useSetRecoilState(editColumnsState);
-  const setDeletedColumns = useSetRecoilState(deletedColumnsState);
-
-  useMemo(() => {
-    if (!isOpen) {
-      setEditColumns(board.columns);
-      setDeletedColumns([]);
-    }
-  }, [board.columns, isOpen, setDeletedColumns, setEditColumns]);
 
   // Session Details
   const { data: session } = useSession({ required: true });
